worker/requirements.txt
opencv-python
timm
einops
torch
torchvision
numpy
requests
pillow
ftfy
regex
transformers
